Add Support for GearScore
Add support for showing GearScore in the Character and Inspect frames. This is shown in addition to average item level.

Showing GearScore is disabled by default to retain the current experience. Players can enable it in the Options if they want to see GearScore. They will have to enable it individually for the Character/Inspect frames -- this is done to retain the current options experience of keeping these settings different.